wip: consolidate pepr things
