+++ package.json
@@ -24,7 +24,8 @@
       "labels": []
     },
     "env": {
-      "UDS_DOMAIN": "###ZARF_VAR_DOMAIN###"
+      "UDS_DOMAIN": "###ZARF_VAR_DOMAIN###",
+      "UDS_WITH_ISTIO": "###ZARF_CONST_WITH_ISTIO###"
     }
   },
   "scripts": {
