--- src/grafana/chart/templates/networkpolicies/ingress-tempo.yaml
@@ -1,23 +1,0 @@
-apiVersion: networking.k8s.io/v1
-kind: NetworkPolicy
-metadata:
-  name: allow-tempo-ingress-writes
-  namespace: {{ .Release.Namespace }}
-spec:
-  podSelector:
-    matchLabels:
-      app: prometheus
-  policyTypes:
-  - Ingress
-  # Allow remote writes from Tempo
-  ingress:
-  - from:
-    - namespaceSelector:
-        matchLabels:
-          app.kubernetes.io/name: tempo
-      podSelector:
-        matchLabels:
-          app.kubernetes.io/name: tempo
-    ports:
-    - port: 9090
-      protocol: TCP
