+++ src/grafana/zarf.yaml
@@ -9,16 +9,16 @@ components:
     required: true
     description: "Deploy Grafana"
     charts:
+      - name: uds-grafana-config
+        namespace: grafana
+        version: 0.1.0
+        localPath: chart
       - name: grafana
         url: https://grafana.github.io/helm-charts/
         version: 7.0.6
         namespace: grafana
         valuesFiles:
           - values/values.yaml
-      - name: uds-grafana-config
-        namespace: grafana
-        version: 0.1.0
-        localPath: chart
     images:
       - docker.io/grafana/grafana:10.1.5
       - docker.io/curlimages/curl:7.85.0
