+++ src/istio/zarf.yaml
@@ -9,6 +9,11 @@ variables:
     description: "Cluster domain"
     default: "uds.dev"
 
+constants:
+  - name: WITH_ISTIO
+    description: "Track istio installation"
+    value: "true"
+
 components:
   - name: istio-controlplane
     description: "Install the Istio control plane based on https://istio.io/latest/docs/setup/install/helm/"
