--- src/loki/chart/templates/networkpolicies/default-deny-all.yaml
@@ -1,12 +1,0 @@
-apiVersion: networking.k8s.io/v1
-kind: NetworkPolicy
-metadata:
-  name: default-deny-all-{{ .Release.Name }}
-  namespace: {{ .Release.Namespace }}
-spec:
-  podSelector: {}
-  policyTypes:
-  - Ingress
-  - Egress
-  egress: []
-  ingress: []
