+++ src/metrics-server/chart/templates/uds-package.yaml
@@ -1,0 +1,5 @@
+apiVersion: uds.dev/v1alpha1
+kind: Package
+metadata:
+  name: metrics-server
+  namespace: { { .Release.Namespace } }
