--- src/neuvector/chart/templates/istio/pepr-config.yaml
@@ -1,9 +1,0 @@
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: uds-neuvector-service-webui
-  namespace: {{ .Release.Namespace }}
-  labels:
-    "uds/istio-gateway": "admin"
-    "uds/istio-host": "neuvector"
-
