+++ src/pepr/config.ts
@@ -1,3 +1,4 @@
 export const UDSConfig = {
   domain: process.env.UDS_DOMAIN || "uds.dev",
+  istioInstalled: process.env.UDS_WITH_ISTIO === "true",
 };
