+++ src/pepr/operator/namespace.ts
