+++ src/pepr/operator/network/allow-egress-within-ns.ts
