--- src/policies/tasks.yaml
@@ -1,4 +1,0 @@
-tasks:
-  - name: validate
-    actions:
-      - cmd: "npx jest src/policies/**/*.spec.ts"
